ccl,keys,sql,storage,util: restructure fuzz tests
