+++ pkg/ccl/pgcryptoccl/pgcryptocipherccl/BUILD.bazel
@@ -4,9 +4,11 @@ go_library(
     name = "pgcryptocipherccl",
     srcs = [
         "cipher.go",
+        "cipher_fuzz.go",
         "cipher_method.go",
         "doc.go",
         "padding.go",
+        "test_helper.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/pgcryptoccl/pgcryptocipherccl",
     visibility = ["//visibility:public"],
@@ -15,6 +17,7 @@ go_library(
         "//pkg/sql/pgwire/pgerror",
         "//pkg/util/errorutil/unimplemented",
         "@com_github_cockroachdb_errors//:errors",
+        "@com_github_stretchr_testify//require",
     ],
 )
