+++ pkg/sql/sqlliveness/slstorage/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "key_encoder.go",
         "metrics.go",
         "sessionid.go",
+        "sessionid_fuzz.go",
         "slstorage.go",
         "test_helpers.go",
     ],
@@ -22,6 +23,7 @@ go_library(
         "//pkg/settings/cluster",
         "//pkg/sql/catalog",
         "//pkg/sql/catalog/systemschema",
+        "//pkg/sql/enum",
         "//pkg/sql/regionliveness",
         "//pkg/sql/sem/eval",
         "//pkg/sql/sqlliveness",
@@ -29,6 +31,7 @@ go_library(
         "//pkg/util/cache",
         "//pkg/util/encoding",
         "//pkg/util/hlc",
+        "//pkg/util/leaktest",
         "//pkg/util/log",
         "//pkg/util/metric",
         "//pkg/util/stop",
@@ -39,6 +42,7 @@ go_library(
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_redact//:redact",
         "@com_github_prometheus_client_model//go",
+        "@com_github_stretchr_testify//require",
     ],
 )
